chore(skill-bill): remediation checkpoint on 'feat/SKILL-132-runtime-kotlin-dead-code-sweep' before mutating-phase re-entry
