+++ runtime-kotlin/ARCHITECTURE.md
@@ -71,7 +71,7 @@ runtime-core
   and the `DecompositionManifestCoherenceValidator`) plus their schema-resource
   copy tasks (`copyInstallPlanSchema`, `copyWorkflowStateSchema`,
   `copyDecompositionManifestSchema`), reached only through domain-neutral ports.
-- `runtime-core`: `RuntimeModule`, Kotlin-Inject component definitions, and DI
+- `runtime-core`: Kotlin-Inject component definitions and DI
   providers. It may know concrete adapters only inside composition code.
   `runtime-core` publishes only the generated Kotlin-Inject ABI edges that its
   public `RuntimeComponent` exposes today: `runtime-application` service types
@@ -1031,8 +1031,7 @@ deferred to subtask 3.
 
 The architecture tests enforce the following rules:
 
-- `ARCHITECTURE.md`, `RuntimeModule.declaredGradleModules`,
-  `RuntimeModule.declaredSubsystemPackages`, Gradle settings, and smoke tests
+- `ARCHITECTURE.md`, Gradle settings, and the architecture-test module catalog
   describe the same module and subsystem graph.
 - `runtime-core` contains only `skillbill` and `skillbill.di` source packages.
 - `runtime-core` does not directly re-export contract or concrete
@@ -1067,8 +1066,6 @@ The architecture tests enforce the following rules:
 - Workflow-state, install-plan, decomposition-manifest, platform-pack,
   native-agent composition, and telemetry-event schema validators are exercised
   at their owning parse seams.
-- Runtime surfaces expose documented `RuntimeSurfaceContract` metadata for
-  active workflow, scaffold, install, native-agent, and launcher operations.
 - typed CLI presenter models are the input to CLI text rendering.
 - `docs/architecture/gradle-module-split-evaluation.md` records the physical
   Gradle split decision and readiness rules.
