--- runtime-kotlin/runtime-application/src/main/kotlin/skillbill/telemetry/config/TelemetryRuntime.kt
@@ -1,18 +1,0 @@
-package skillbill.telemetry.config
-
-import skillbill.application.telemetry.loadTelemetrySettings
-import skillbill.ports.telemetry.TelemetrySettingsProvider
-import skillbill.telemetry.model.TelemetrySettings
-
-object TelemetryRuntime {
-  fun loadTelemetrySettings(
-    settingsProvider: TelemetrySettingsProvider,
-    materialize: Boolean = false,
-  ): TelemetrySettings = TelemetryConfigRuntime.loadTelemetrySettings(
-    materialize = materialize,
-    settingsProvider = settingsProvider,
-  )
-
-  fun telemetryIsEnabled(settingsProvider: TelemetrySettingsProvider): Boolean =
-    TelemetryConfigRuntime.telemetryIsEnabled(settingsProvider)
-}
