+++ runtime-kotlin/runtime-application/src/test/kotlin/skillbill/workflow/NoopGoalPlanningPreparationEnvelopeValidator.kt
@@ -1,0 +1,5 @@
+package skillbill.workflow
+
+object NoopGoalPlanningPreparationEnvelopeValidator : GoalPlanningPreparationEnvelopeValidator {
+  override fun validate(envelope: Map<String, Any?>, sourceLabel: String) = Unit
+}
