+++ runtime-kotlin/runtime-cli/build.gradle.kts
@@ -28,10 +28,9 @@ dependencies {
   runtimeOnly(libs.slf4j.nop)
   ksp(libs.kotlin.inject.compiler)
   // SKILL-52.2 subtask 5: runtime-infra-fs / runtime-infra-http stay on the
-  // test classpath because `RuntimeModuleSmokeTest` and adapter-side smoke tests
-  // import concrete runtime classes (`InstallRuntime`, `LauncherRuntime`,
-  // `NativeAgentRuntime`, `ScaffoldRuntime`, telemetry HTTP runtime). Test code
-  // crossing module boundaries for fixtures is expected; main source must not.
+  // test classpath because adapter-side smoke tests import concrete runtime
+  // classes. Test code crossing module boundaries for fixtures is expected;
+  // main source must not.
   testImplementation(project(":runtime-infra-fs"))
   testImplementation(project(":runtime-infra-http"))
   testImplementation(project(":runtime-infra-sqlite"))
