--- runtime-kotlin/runtime-contracts/src/main/kotlin/skillbill/contracts/ContractEnvelope.kt
@@ -1,7 +1,0 @@
-package skillbill.contracts
-
-/** Generic payload wrapper for contract-bound runtime interactions. */
-data class ContractEnvelope<T : Any>(
-  val contract: RuntimeContract,
-  val payload: T,
-)
