+++ runtime-kotlin/runtime-core/build.gradle.kts
@@ -21,6 +21,7 @@ dependencies {
   // SKILL-129 subtask 5: the durable/telemetry redaction proof drives the production review runner
   // through the shared recording harness rather than hand-building an accounting summary.
   testImplementation(testFixtures(project(":runtime-application")))
+  testImplementation(testFixtures(project(":runtime-ports")))
   testImplementation(libs.junit.jupiter)
   testImplementation(libs.kotlin.test)
 }
