+++ runtime-kotlin/runtime-domain/src/main/kotlin/skillbill/workflow/GoalPlanningPreparationEnvelopeValidator.kt
@@ -6,7 +6,3 @@ interface GoalPlanningPreparationEnvelopeValidator {
   @OpenBoundaryMap("Goal planning preparation wire map at the schema-validation seam")
   fun validate(envelope: Map<String, Any?>, sourceLabel: String)
 }
-
-object NoopGoalPlanningPreparationEnvelopeValidator : GoalPlanningPreparationEnvelopeValidator {
-  override fun validate(envelope: Map<String, Any?>, sourceLabel: String) = Unit
-}
