--- runtime-kotlin/runtime-infra-fs/src/main/kotlin/skillbill/install/runtime/InstallRuntime.kt
@@ -1,40 +1,0 @@
-package skillbill.install.runtime
-
-import skillbill.contracts.surface.RuntimeSurfaceContract
-import skillbill.contracts.surface.RuntimeSurfaceStatus
-
-/** Kotlin-owned install primitives used by scaffold and CLI adapter surfaces. */
-object InstallRuntime {
-  val contract: RuntimeSurfaceContract = RuntimeSurfaceContract(
-    name = "install",
-    ownerPackage = "skillbill.install",
-    contractVersion = "0.1",
-    status = RuntimeSurfaceStatus.ACTIVE,
-    summary = "Agent-path detection, skill symlink installation, and install rollback primitives.",
-    placeholderReason = "",
-    supportedOperations = listOf(
-      "plan-install",
-      "apply-install",
-      "agent-path",
-      "detect-agents",
-      "codex-agents-path",
-      "claude-agents-path",
-      "opencode-agents-path",
-      "junie-agents-path",
-      "cursor-agents-path",
-      "link-skill",
-      "cleanup-agent-target",
-      "link-claude-agents",
-      "unlink-claude-agents",
-      "link-codex-agents",
-      "unlink-codex-agents",
-      "link-opencode-agents",
-      "unlink-opencode-agents",
-      "link-junie-agents",
-      "unlink-junie-agents",
-      "link-cursor-agents",
-      "unlink-cursor-agents",
-      "rollback-links",
-    ),
-  )
-}
