+++ runtime-kotlin/runtime-infra-fs/src/main/kotlin/skillbill/scaffold/rendering/ScaffoldTemplateRendering.kt
@@ -121,17 +121,6 @@ internal fun renderContentBody(
   }
 }
 
-internal fun renderSkillBody(context: TemplateContext, description: String, areaFocus: String = ""): String =
-  buildString {
-    append(renderFrontmatter(context.skillName, description))
-    appendLine()
-    append(renderDescriptorSection(context, areaFocus))
-    appendLine()
-    append(CANONICAL_EXECUTION_SECTION)
-    appendLine()
-    append(renderCeremonySection(skillClass = null))
-  }
-
 internal fun renderAddonBody(skillName: String, description: String, explicitBody: String?): String {
   val body = explicitBody ?: buildString {
     appendLine("# $skillName")
