+++ runtime-kotlin/runtime-ports/build.gradle.kts
@@ -1,6 +1,10 @@
 plugins {
   id("skillbill.jvm-library")
   id("skillbill.quality")
+  // SKILL-132 subtask 1: the empty port implementations are test-only fakes consumed by
+  // runtime-core and runtime-application tests, so they ship as published test fixtures
+  // instead of production declarations.
+  `java-test-fixtures`
 }
 
 dependencies {
