+++ runtime-kotlin/runtime-ports/src/main/kotlin/skillbill/ports/persistence/WorkListRepository.kt
@@ -5,7 +5,3 @@ import skillbill.ports.persistence.model.WorkItem
 interface WorkListRepository {
   fun list(limit: Int? = null): List<WorkItem>
 }
-
-object EmptyWorkListRepository : WorkListRepository {
-  override fun list(limit: Int?): List<WorkItem> = emptyList()
-}
