+++ runtime-kotlin/runtime-ports/src/main/kotlin/skillbill/ports/review/ReviewAttributionPort.kt
@@ -3,7 +3,3 @@ package skillbill.ports.review
 interface ReviewAttributionPort {
   fun routedSkillPlatformSlugs(): Map<String, String>
 }
-
-object EmptyReviewAttributionPort : ReviewAttributionPort {
-  override fun routedSkillPlatformSlugs(): Map<String, String> = emptyMap()
-}
