+++ runtime-kotlin/runtime-ports/src/testFixtures/kotlin/skillbill/ports/persistence/EmptyWorkListRepository.kt
@@ -1,0 +1,7 @@
+package skillbill.ports.persistence
+
+import skillbill.ports.persistence.model.WorkItem
+
+object EmptyWorkListRepository : WorkListRepository {
+  override fun list(limit: Int?): List<WorkItem> = emptyList()
+}
