+++ runtime-kotlin/runtime-ports/src/testFixtures/kotlin/skillbill/ports/review/EmptyReviewAttributionPort.kt
@@ -1,0 +1,5 @@
+package skillbill.ports.review
+
+object EmptyReviewAttributionPort : ReviewAttributionPort {
+  override fun routedSkillPlatformSlugs(): Map<String, String> = emptyMap()
+}
